feat: overlapping tag handling
Added handling of overlapping tags. Currently, it is assumed that overlapping tags are a rare occurrence (as they are likely just artifacts due to unintentional sequence similarity). From this follows that, for most cases, given a set of tags, it is possible to find a maximum non-overlapping subset by removing just one tag. More complicated cases are currently discarded, as they are very rare. Proper logic to handle those cases might be added later.