enh: iterate reads using noodles
Replaced manual implementation of the read iterator with the one from noodles. This should simplify an eventual transition to oxbow, as well as making compression handling easier (though currently output is not being compressed). Since read sequence becomes a vector of bytes, adjusted functions accordingly.

Also made so that BarcodeBuilder works with references to avoid copying the Tag objects.