feat: added overlap removal logging
If a barcode underwent any form of tag overlap removal, log the result in a tsv log file.
The first column of the log file contains the fixed barcode while the second one contains the removed tags.
Still a bit rough around the edges, might need a bit of polish down the line.