enh: improved tag formatting in the output file
Previously, only the name of the tag was saved in the read header, losing information about the actual tag sequence (and thus length) and relying on the name of the tag to infer information about it (such as tag type). This was not ideal, as it required dragging along the tag file to reconstruct full information and plot tag alignment.

Changed so that now tag sequence, info and alignment position are saved for each tag (separated by ":"). Also took the opportunity to replace the tag separator to "|" instead of the previous "[]", which made parsing extremely tedious.